requirements.txt
guizero>=1.1.0
mplayer.py>=0.7.1
paho-mqtt>=1.3.1
pyaml>=17.12.1
PyYAML>=5.1.2
